flea-tools 优化“JPA代码生成器”工具，支持实体类使用lombok
