+++ flea-tools/src/main/resources/code.properties
@@ -1,14 +1,16 @@
-dbName=fleaauth
-tableName=flea_resource
-tableNameDesc=Flea\u8D44\u6E90
+dbName=fleafs
+tableName=flea_file_version
+tableNameDesc=Flea\u6587\u4EF6\u7248\u672C
 idGeneratorTable=flea_id_generator
 pkColumnName=id_generator_key
 valueColumnName=id_generator_value
 author=huazie
-version=2.0.0
-rootPath=E:\\fleaworkspace\\flea-framework\\flea-auth\\src\\main\\java
-codePackage=com.huazie.fleaframework.auth.base.function
-puDaoCodePackage=com.huazie.fleaframework.auth.base
-puDaoClassName=FleaAuthDAOImpl
-puName=fleaauth
-puAliasName=FleaAuth
+version=1.0.0
+rootPath=E:\\fleaworkspace\\FleaFS\\fleafs-repository\\src\\main\\java
+#rootPath=E:\\fleaworkspace\\flea-jpa-test\\src\\main\\java
+#rootPath=F:\\FileRecv\\project\\flea-fleaframework\\flea-core\\src\\main\\java
+codePackage=com.huazie.ffs.base
+puDaoCodePackage=com.huazie.ffs.base.dao.impl
+puDaoClassName=FleaFSDAOImpl
+puName=fleafs
+puAliasName=FleaFS
