+++ flea-tools/src/main/resources/flea/code/entity/FleaEntityWithLombok.code
@@ -1,0 +1,36 @@
+package %CODE_PACKAGE%.entity;
+
+import com.huazie.fleaframework.common.FleaEntity;
+import lombok.Getter;
+import lombok.Setter;
+import lombok.ToString;
+
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.SequenceGenerator;
+import javax.persistence.Table;
+import javax.persistence.TableGenerator;
+import javax.persistence.Temporal;
+import javax.persistence.TemporalType;
+import java.util.Date;
+
+/**
+ * %TABLE_DESC%表对应的实体类
+ *
+ * @author %AUTHOR%
+ * @version %VERSION%
+ * @since %VERSION%
+ */
+@Getter
+@Setter
+@ToString
+@Entity
+@Table(name = "%TABLE_NAME%")
+public class %ENTITY_CLASS_NAME% extends FleaEntity {
+
+%VARIABLE_CODE%
+
+}
