+++ flea-tools/src/main/resources/flea/i18n/flea_i18n_common.properties
@@ -45,6 +45,7 @@ COMMON_CODE_00022=\u4E3B\u952E\u751F\u6210\u7B56\u7565
 COMMON_CODE_00023=\u4E3B\u952E\u751F\u6210\u5668\u8868
 COMMON_CODE_00024=\u4E3B\u952E\u5217\u540D
 COMMON_CODE_00025=\u4E3B\u952E\u503C\u5217\u540D
+COMMON_CODE_00026=\u5B9E\u4F53\u7C7B\u4F7F\u7528Lombok
 #
 COMMON_TOOLS_00000=\u6587\u4EF6
 COMMON_TOOLS_00001=\u9000\u51FA
