+++ flea-tools/src/main/resources/flea/i18n/flea_i18n_common_en_US.properties
@@ -45,6 +45,7 @@ COMMON_CODE_00022=ID Generator Strategy
 COMMON_CODE_00023=ID Generator Table
 COMMON_CODE_00024=PK Column Name
 COMMON_CODE_00025=Value Column Name
+COMMON_CODE_00026=EntityClass With Lombok
 #
 COMMON_TOOLS_00000=File
 COMMON_TOOLS_00001=Exit
