+++ flea-tools/src/test/java/com/huazie/fleaframework/tools/FleaToolsTest.java
@@ -1,13 +1,12 @@
 package com.huazie.fleaframework.tools;
 
 /**
- * <p></p>
  *
  * @author huazie
  * @version 1.0.0
  * @since 1.0.0
  */
-public class FleaFrameToolsTest {
+public class FleaToolsTest {
 
     public static void main(String[] args){
         StartMain.main(args);
